feat: Add card generation and larger board
This is the start of adding more features to the game. Started with
a card generation function and a shuffle function so that a randomized
board can be made. This will then turn into a game that can be reset and
the size of the board customized by the user. There is a lot that needs
changed before it can be considered production.